+++ hkl/hkl-pseudoaxis-common-eulerians.c
@@ -80,7 +80,7 @@ static int hkl_pseudo_axis_engine_mode_get_eulerians_real(HklPseudoAxisEngineMod
 
 	hkl_geometry_update(geometry);
 
-	solution = self->parameters[0].value;
+	solution = (int)self->parameters[0].value;
 
 	komega = ((HklParameter *)hkl_geometry_get_axis_by_name(geometry, "komega"))->value;
 	kappa = ((HklParameter *)hkl_geometry_get_axis_by_name(geometry, "kappa"))->value;
@@ -105,7 +105,7 @@ static int hkl_pseudo_axis_engine_mode_set_eulerians_real(HklPseudoAxisEngineMod
 
 	double angles[3];
 
-	solution = self->parameters[0].value;
+	solution = (int)self->parameters[0].value;
 
 	status |= eulerian_to_kappa(((HklParameter *)engine->pseudoAxes[0])->value,
 				    ((HklParameter *)engine->pseudoAxes[1])->value,
